Update setup version Rename .mvomrc.json to manifest.json
