+++ src/manifest.json
@@ -9,7 +9,7 @@
     "readFileContentsById": "^1.0.0",
     "getServerInfo": "^1.1.0",
     "save": "^2.0.0",
-    "setup": "^1.0.0",
+    "setup": "^1.1.0",
     "teardown": "^1.0.0"
   }
 }
